+++ base/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ahmad.cv.base">
+    package="com.faizan_atif.cv.base">
 
     <uses-permission android:name="android.permission.CALL_PHONE" />
     <application
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.ahmad.cv.activities.ScrollingActivity"
+            android:name="com.faizan_atif.cv.activities.ScrollingActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
