+++ base/src/main/java/com/faizan_atif/cv/adapters/EducationAdapter.java
@@ -1,4 +1,4 @@
-package com.ahmad.cv.adapters;
+package com.faizan_atif.cv.adapters;
 
 import android.app.Activity;
 import android.content.Context;
@@ -13,8 +13,8 @@ import android.widget.ImageView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.ahmad.cv.model.DemoItem;
-import com.ahmad.cv.base.R;
+import com.faizan_atif.cv.model.DemoItem;
+import com.faizan_atif.cv.base.R;
 
 import java.util.List;
 public class EducationAdapter extends RecyclerView.Adapter {
