+++ base/src/main/java/com/faizan_atif/cv/adapters/ViewPagerAdapter.java
@@ -1,4 +1,4 @@
-package com.ahmad.cv.adapters;
+package com.faizan_atif.cv.adapters;
 
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
