+++ base/src/main/java/com/faizan_atif/cv/fragments/ProfessionFragment.java
@@ -1,28 +1,29 @@
-package com.ahmad.cv.fragments;
+package com.faizan_atif.cv.fragments;
 import android.os.Bundle;
+
+import com.faizan_atif.cv.model.DemoItem;
+import com.faizan_atif.cv.adapters.ProfessionAdapter;
+import com.faizan_atif.cv.base.R;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-
-import com.ahmad.cv.model.DemoItem;
-import com.ahmad.cv.adapters.EducationAdapter;
-import com.ahmad.cv.base.R;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import java.util.ArrayList;
 import java.util.List;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class EducationFragment extends Fragment {
+public class ProfessionFragment extends Fragment {
 
-    public EducationFragment() {
+    public ProfessionFragment() {
         // Required empty public constructor
     }
-    public static EducationFragment newInstance() {
-        EducationFragment fragment = new EducationFragment();
+    public static ProfessionFragment newInstance() {
+        ProfessionFragment fragment = new ProfessionFragment();
         Bundle args = new Bundle();
         fragment.setArguments(args);
         return fragment;
@@ -35,14 +36,15 @@ public class EducationFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_blank_fragment2, container, false);
-        RecyclerView recyclerViewDemo = view.findViewById(R.id.recyclerViewDemo1);
+        View view = inflater.inflate(R.layout.fragment_blank_fragment3, container, false);
+        RecyclerView recyclerViewDemo = view.findViewById(R.id.recyclerViewDemo3);
         recyclerViewDemo.setLayoutManager(new LinearLayoutManager(getContext()));
-       recyclerViewDemo.setAdapter(new EducationAdapter(feedItems(), getContext()));
+        recyclerViewDemo.setAdapter(new ProfessionAdapter(feedItems(), getContext()));
         return view;
     }
+
     private List<DemoItem> feedItems() {
-        String Titles = "Education Section" ;
+        String Titles = "Profession Section";
         List<DemoItem> demoItems = new ArrayList<>();
         DemoItem demoItem = new DemoItem(Titles);
         demoItems.add(demoItem);
@@ -50,7 +52,3 @@ public class EducationFragment extends Fragment {
 
     }
 }
-
-
-
-
