+++ base/src/main/java/com/faizan_atif/cv/model/DemoItem.java
@@ -1,4 +1,4 @@
-package com.ahmad.cv.model;
+package com.faizan_atif.cv.model;
 
 public class DemoItem {
     private String title;
