+++ base/src/main/res/menu/menu_scrolling.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.ahmad.cv.activities.ScrollingActivity">
+    tools:context="com.faizan_atif.cv.activities.ScrollingActivity">
     <item
         android:id="@+id/action_settings"
         android:title="@string/action_settings"
