+++ installed/build.gradle
@@ -6,7 +6,7 @@ android {
 
 
     defaultConfig {
-        applicationId "com.ahmad.resume"
+        applicationId "com.faizan_atif.resume"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
