+++ installed/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.ahmad.coordinatetabs">
+    package="com.faizan_atif.coordinatetabs">
 
     <application
         android:allowBackup="true"
