+++ installed/src/test/java/com/solo9/coordinatetabs/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.ahmad.coordinatetabs;
+package com.faizan_atif.coordinatetabs;
 
 import org.junit.Test;
